After merge code-style fix
